requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-storage
pymongo
python-dotenv
